Move `MacroKind` into `libsyntax_pos`
So it can be eventually used in `ExpnInfo`
